+++ devtools/android/src/main/kotlin/com/maximbircu/devtools/android/presentation/tools/enum/EnumToolLayout.kt
@@ -1,18 +1,18 @@
 package com.maximbircu.devtools.android.presentation.tools.enum
 
 import android.content.Context
+import android.view.View
+import android.widget.TextView
 import com.maximbircu.devtools.android.R
-import com.maximbircu.devtools.android.extensions.hide
 import com.maximbircu.devtools.android.extensions.setOnClickListener
-import com.maximbircu.devtools.android.extensions.show
 import com.maximbircu.devtools.android.presentation.tool.DevToolLayout
+import com.maximbircu.devtools.android.presentation.tools.enum.selectors.chips.EnumToolChipsOptionSelectorLayout
 import com.maximbircu.devtools.android.presentation.tools.enum.selectors.dialog.EnumToolOptionSelectorDialog
 import com.maximbircu.devtools.common.presentation.tools.enum.EnumTool
 import com.maximbircu.devtools.common.presentation.tools.enum.EnumToolPresenter
 import com.maximbircu.devtools.common.presentation.tools.enum.EnumToolView
 import kotlinx.android.synthetic.main.layout_dev_tool.view.devToolCard
-import kotlinx.android.synthetic.main.layout_enum_tool.view.chipsSelector
-import kotlinx.android.synthetic.main.layout_enum_tool.view.currentValue
+import kotlinx.android.synthetic.main.layout_enum_tool.view.contentContainer
 
 class EnumToolLayout(context: Context) : DevToolLayout<EnumTool>(context), EnumToolView {
     private val presenter = EnumToolPresenter.create(this)
@@ -21,18 +21,12 @@ class EnumToolLayout(context: Context) : DevToolLayout<EnumTool>(context), EnumT
     override fun onBind(tool: EnumTool) = presenter.onToolBind(tool)
 
     override fun showCompactOptionsSelector(tool: EnumTool, onNewOptionSelected: (String) -> Unit) {
-        chipsSelector.show()
-        chipsSelector.bind(tool, onNewOptionSelected)
+        setContentView(EnumToolChipsOptionSelectorLayout(context, tool, onNewOptionSelected))
     }
 
     override fun showConfigurationValue(value: String) {
-        currentValue.show()
-        currentValue.text = value
+        setContentView(TextView(context).apply { text = value })
         devToolCard.setOnClickListener(presenter::onToolClick)
-    }
-
-    override fun hideConfigurationValue() {
-        currentValue.hide()
     }
 
     override fun showOptionSelectorDialog(tool: EnumTool, onNewOptionSelected: (String) -> Unit) {
@@ -45,8 +39,8 @@ class EnumToolLayout(context: Context) : DevToolLayout<EnumTool>(context), EnumT
         ).show()
     }
 
-    override fun hideCompactOptionsSelector() {
-        chipsSelector.hide()
-        devToolCard.setOnClickListener(null)
+    private fun setContentView(contentView: View) {
+        contentContainer.removeAllViews()
+        contentContainer.addView(contentView)
     }
 }